i have no idea what vert.txt is, it keeps cropping up as output from something.
adding loops to generate null models, but i did a stupid thing and generated it for a statistics i don't need a null model for.  need to swap in species-species distances and run the first loop, or see if the second loop works and test each community against its null self.